refactor: PostAuthors folder
