+++ src/layouts/PostLayout/PostLayout.tsx
@@ -8,7 +8,7 @@ import ScrollTopAndComment from 'src/components/ScrollTopAndComment'
 import SectionContainer from 'src/components/SectionContainer'
 import Tag from 'src/components/Tag'
 import { NextPost } from './NextPost'
-import { PostAuthors } from './PostAuthors'
+import { PostAuthors } from './PostAuthors/PostAuthors'
 import { PostHeader } from './PostHeader'
 import { PreviousPost } from './PreviousPost'
 import { TableOfContents } from './TableOfContents/TableOfContents'
